Add FCB logo and info
Futbol Club (FCB) is a BSC project of coin an ecological chain. It is a global football sports prediction machine. It takes the lead in the world in the way of ecological mining pool (EPM) to produce. It participates in project ecology, that is, mining. By participating in FCB ecology, all contributors can obtain FCB awards, and FCB can participate in various sports events, With the unchangeable characteristics of blockchain, it enhances the fairness of all kinds of sports events, and creates an open, inclusive and safe global competitive sports ecosystem. Meanwhile, FCB is committed to creating and improving a highly autonomous community based on blockchain technology and digital assets, realizing the leap of blockchain 3.0 era.