Add NB logo and info
NB.Token coinage is an improved token of many similar mechanisms, such as inflation, cross-link technology compatibility to reduce GAS costs, algorithm upgrade, landing applications and so on. NB.T oken is a completely decentralized community experiment. Because its price is very low, the appreciation is empty to the huge, deeply loves by the vast block chain enthusiast. NB Niu B will create multi-chain + cross-connection + aggregation as a technical feature, using the extreme deflation mechanism, 50% of the total will be handed over to the black hole address for combustion, and destroy the key, each transfer involved in the transaction will be deducted 5% tax, of which 3% will be allocated to the participants in the fund pool, 1% will be destroyed by themselves, 1% will be prorated to the local currency holder account, and the NB community will be completely autonomous by the currency holder.