Add C19VT info and logo
Token name:  Covid-19 Vaccine Token
Token symbol:  C19VT
Total supply:  1,000 trillion

NFT+MEME concept on the Binance Smart Chain. Jointly initiated by Silicon Valley star company HealthFi and School of Medicine, Johns Hopkins University. Created for fund management in the field of health and medical therapy, striving to pay for the research and vaccination costs of the Covid-19 vaccine in the future. 

Greatest respect to the medical staff who dedicated to fight against the pandemic in the front and deeply mourn to the hundreds of thousands of deaths in the past two years. There is a better future with Covid-19 Vaccine and its tokens.

40% of the tokens are produced by liquidity mining, 30% are put into the initial liquidity pool, 25% are locked-up with 1% releasing every month after 1 year, remaining 5% airdrop rewards for global community. No ownership, 100% decentralized autonomy. 

Launch time:
13:00(GMT); Delhi 18:30(GMT+5:30); Singapore 21:00(GMT+8); Washington 9:00(GMT-5), 26th May 2021.

BNB/C19VT is now available on PancakeSwap.

Contract address:
0x4064F3E26b9C547D13882A56d50aCb895dCc95cb