Add files via upload
JKDAO, with a total supply of 1,000 coins  The loss-free staking of both JOKER and DFS is supported for the first version.  The JOKER POOL produces 1 JKDAO per day.  The DFS Pool produces 0.2 JKDAO per day.   JKDAO is the governance token of JOKER.ONE, and the participants enjoy 50%  of fee.joker’s dividend, as well as the eligibility of proposal.  Holding 1 JKDAO enjoys the dividend right.  The dividend will be randomly snapshotted on a weekly basis and airdropped directly to the accounts.