Bananashark（BBS）
BananaShark is a completely decentralized TOKEN driven by DOGE and SHIB global community, which will be a decentralized experiment.The total amount of BBS is 1,000 trillion dollars.10% tokens will be airdropped, 40% tokens will be directly into the black hole for destruction, and the remaining 50% will be put into the liquidity pool. The project party has no token reserve; everyone has to make purchase at Pancakeswap. It means that the project party has no more rights than the users have. The poll is permanently locked, and more activities like Pledge mining and NFT will be turn on in the future.
The website is bbscoin.co
Email:BananaShark@gmail. com