Add MDOGE loge and info
MDOGE was established in Silicon Valley, USA in March 2021, headquartered in Singapore, and is composed of a team of top international blockchain geeks.
The founding members of the MDOGE team come from world-renowned technology companies from Microsoft and Google. They have an average of more than 3 years of technical development experience. They have rich work experience and a number of mature Project experience. At present, most of the core members of the team are located in Singapore and San Francisco.