add FUWA info and logo
1000TPT交易链接ID:0xdbb5b3d0e077659aaec53b51a3787ce4a968da93a8e55c840fa1b8e51e780d74

FUWA简介：Fuwa represents good luck. Every Olympic Games will have the birth of Fuwa mascot. The coin we issued with Fuwa as the theme is to commemorate the meaningful Olympic Games, and write it into the blockchain for permanent collection. We will use the unique three-dimensional economic model, token ➕ defi ➕ Three dimensional integration of NFT to create a unique ecosystem