XOM
Pulsar DPAAS:  a decentralized development platform as a service (DPAAS）,which combines the current mature SaaS and PaaS technologies to drive the development of with its own advantages. It creates a distributed open source of the blockchain cloud platform with the underlying technologies, which will evolve into a highly redundant incentive peer-to-peer distributed environment. It is a highly decentralized application, a new original financial ecology, and a new interoperable DPAAS system.