Add TEA logo and info
Teachain completely relies on the physical business model. We have reached cooperation with nearly ten well-known tea companies and will provide them with traceability services, laying traceability equipment for their tea mountains, processing plants, and tea products, and well-known teas that have obtained traceability services. Sell in the tea chain mall, and cooperate with tea companies, well-known platforms in the industry, and a large number of offline distributors for promotion. We purchase goods from tea companies at low cost, and rebate consumers in the form of points. The intermediate profit is the profit of the mall. The financial of the mall will be announced on a periodic basis, and the tokens will be repurchased with this profit. The plan and execution results will be announced accordingly.
Then use TEA's innovative model to penetrate into the ecology of the tea industry. Based on a large amount of accurate traffic in the tea industry, launch such as tea credits and tea auctions to give TEA use value. In the future, we will implement the mode of holding currency to vote for new tea. All tea companies need to vote for new tea in the mall. This is the core part of TEA's needs.
https://cn.etherscan.com/tx/0xefd9d5d8593814e42326f62a838f8066fc9f3e8ccec1490d3fc6477c3ab180db