Add files via upload
Unigains aims to create a token that combines a whole new generation of tokenomics, combined with the original “BuyBack Wallet” methodology. This newly founded generation of tokenomics will create everlasting gains for the project itself and to the community as a whole.

The developers of UniGains elected to use an updated and revised version of the first original contract that deployed the “BuyBack Smart Chain Methodology”. The internal mechanics of the coin, as well as the fee structure includes allocating a portion of the collected fees to reinvest back into the token. This mechanism of tokenomics is tested and has proven to be a game changer for both investors and the project itself. UniGains aspires to incorporate three very powerful tools within the BSC cryptocurrency space: combining a dynamically enhanced tokenomic system, superior marketing, and a continually growing community.

UniGains is the first token to form an Alliance with "BuyBack Token (BBT)," by pairing a portion of initial liquidity with the original BuyBack Token (BBT).

0xcee08632fad3e762c49143559e2a8280385612b28727113c9106b0e7f5095993