Add TCM logo and info
TCM is a hyper-deflationary DeFi token that runs on the Binance Smart Link. TCM uses burning and reflection mechanisms, with sign-in mining functions, allowing more users to sign-in mining Own tokens. In the later stage of the project, card games and NFT will be launched, and a new diversified decentralized trading network will be established.
https://bscscan.com/tx/0x33efbe4d0d5a908f17efeacaa22a8c375a338a59e480c940c0a29576f3c09017