Add files via upload
On May 8, 1886, Coca Cola was born in Atlanta, Georgia. It brings people around the world a wonderful feeling of happiness every day! On December 8, 2013, dogecoin was born and won the hearts of the people with the culture of fairness, charity and reward! Now, what kind of innovative inspiration will they inspire when they blend with each other?
https://bscscan.com/tx/0xa19efe30e7de92cd6bae2057804306ad98af0d67e0f91c42a4ce4178e4243b7a