ADD KingFeg info and logo
KINGFEG is a juvenile orangutan in the circus. It lives in the background of the battle among the blockchain empires. The city was destroyed and the social order was in chaos. The escaped KINGFEG inadvertently hid in Dr. Franklin’s research institute. It was the first time it came into contact with the blockchain and was surprised to find that it could vaguely understand it, so it was immersed in these wonderful symbols. After a long period of time, it has mastered the mystery of the blockchain and surprisingly built its own set of simple blockchain systems. It looked for the way to circus, went back there, and found the rest of the mates. Start to teach them blockchain knowledge in their language. A blockchain empire that transcends species is rising. Gradually, other blockchain empires realized this already quietly powerful empire, so the war was about to start. The process was cruel and long. KINGFEG and its partners continued to grow in frustration, suffering, and joy. In the end, they swept away and occupied the top of the blockchain, becoming the creator of the new era.
Transaction ID link：https://bscscan.com/tx/0x42b3353aee6092900525e594471bfc2b68212601e2889caa98217286ba8d82e7