ADD HDD Logo  and info
HDD is a token of sesame trading platform, which is initiated by HD foundation. The profits generated by users through trading strategy are issued in the form of HDD pass. HDD adopts Shib distribution model, which is widely used in the fields of asset management, trading, circulation, payment and settlement of sesame trading platform. With the international head trading platform as the main trading channel, HDD has a strong value-added space.