Add AK logo and info
The Ant Foundation of Singapore and the Ant Community will develop the Ant King English (Ant King) token AK in the Defi field on the BSC chain in January 2022, with a total circulation of 3,888 pieces, and will never be issued more. The Ant Foundation of Singapore developed an ant business system, similar in style to Meituan, Didi Taxi, Taobao, etc. And we have a real ABS consumption value-added system, consumers can consume value-added online and offline stores, so that all consumers can consume without pain.
https://bscscan.com/tx/0x705d885b9b29090888fad2aee74424b1e6d8bda5594bdebcecfab21a75644f0c
https://bscscan.com/tx/0x59e67f109435775c72691f200dbba2a0b2bfa103af7097cd88c5f9d37f25ca91